1. split verification tests for each pool 2. move common functions to verification-utils.js 3. add verification tests of polygon 4. add error check in shell scripts
